requirements.txt
allure-pytest==2.8.18
pytest-testconfig==0.2.0
selenium==3.141.0